requirements.txt
google-genai
PyMuPDF
PyYAML
pydantic
rapidfuzz
spacy
tqdm
python-dotenv